[ttx_diff] Split up normalizer gpos module
